+++ react/src/router/Router.jsx
@@ -1,14 +1,14 @@
 import React from 'react'
 import { Navigate, Route, Routes } from "react-router-dom";
 
-import LoginPage from '../views/auth/Login'
-import PageNotFound from '../pages/PageNotFound'
+import Login from '../pages/auth/Login'
+import PageNotFound from '../pages/404/PageNotFound'
 
 export const Router = () => {
 
 return (
     <Routes>
-        <Route path="/" element={<LoginPage />} />
+        <Route path="/" element={<Login />} />
         <Route path="/*" element={<PageNotFound />} />
     </Routes>
   );
